refactor(webapp-game): move off createProject move to actionMoves
