+++ packages/game/src/index.ts
@@ -9,6 +9,7 @@ import forceCards from './data/card/forces.json';
 import eventCards from './data/card/events.json';
 import { isInRange } from './utils';
 import { ActiveProject, ActiveProjects } from './activeProjects';
+import { createProject } from './moves/actionMoves';
 
 type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 type WithGameState<G extends any, F extends (...args: any) => void> = (G: State<G>['G'], ctx: State<G>['ctx'], ...args: Parameters<F>) => any;
@@ -116,67 +117,8 @@ export const OpenStarTerVillage: Game<type.State.Root> = {
       action: {
         moves: {
           createProject: {
-            // client cannot see decks, discard job card should evaluated on server side
             client: false,
-            move: ((G, ctx, projectCardIndex, jobCardIndex) => {
-              if (!G.table.activeMoves.createProject) {
-                return INVALID_MOVE;
-              }
-
-              const currentPlayer = ctx.playerID!;
-              const currentPlayerToken = G.players[currentPlayer].token;
-              // TODO: replace hardcoded number with dynamic rules
-              const createProjectActionCosts = 2;
-              if (currentPlayerToken.actions < createProjectActionCosts) {
-                return INVALID_MOVE;
-              }
-              const createProjectWorkerCosts = 1;
-              if (currentPlayerToken.workers < createProjectWorkerCosts) {
-                return INVALID_MOVE;
-              }
-
-              // check project card in in hand
-              const currentHandProjects = G.players[currentPlayer].hand.projects;
-              if (!isInRange(projectCardIndex, currentHandProjects.length)) {
-                return INVALID_MOVE;
-              }
-
-              // check job card is on the table
-              const currentJobs = G.table.activeJobs;
-              if (!isInRange(jobCardIndex, currentJobs.length)) {
-                return INVALID_MOVE;
-              }
-
-              // check job card is required in project
-              const projectCard = Cards.GetById(currentHandProjects, projectCardIndex);
-              const jobCard = Cards.GetById(currentJobs, jobCardIndex);
-              if (!Object.keys(projectCard.requirements).includes(jobCard.name)) {
-                return INVALID_MOVE;
-              }
-
-              // reduce action tokens
-              currentPlayerToken.actions -= createProjectActionCosts;
-              Cards.RemoveOne(currentHandProjects, projectCard);
-              Cards.RemoveOne(currentJobs, jobCard);
-
-              // initial active project
-              const activeProjectIndex = ActiveProjects.Add(G.table.activeProjects, projectCard, currentPlayer);
-              const activeProject = ActiveProjects.GetById(G.table.activeProjects, activeProjectIndex);
-
-              // reduce worker token
-              currentPlayerToken.workers -= createProjectWorkerCosts;
-              // assign worker token
-              const jobInitPoints = 1;
-              ActiveProject.AssignWorker(activeProject, jobCard.name, currentPlayer, jobInitPoints);
-              // score victory points
-              const createProjectVictoryPoints = 1;
-              G.players[currentPlayer].victoryPoints += createProjectVictoryPoints;
-
-              // discard job card
-              Deck.Discard(G.decks.jobs, [jobCard]);
-
-              G.table.activeMoves.createProject = false;
-            }) as WithGameState<type.State.Root, type.Move.CreateProject>,
+            move: createProject,
           },
           recruit: {
             // client cannot see decks, discard job card should evaluated on server side
